backend/requirements.txt
Django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
Pillow==10.2.0
numpy==1.26.4
SpeechRecognition==3.10.1
djangorestframework-simplejwt==5.3.1 